int batch size

+++ personality_captions/run_pic.py
@@ -450,7 +450,7 @@ def validate(args, model, tokenizer):
     val_dataset = build_dataset(op.join(args.data_dir, args.val_yaml), tokenizer, args, is_train=True)
     val_sampler = SequentialSampler(val_dataset)
     train_dataloader = DataLoader(val_dataset, sampler=val_sampler,
-                                  batch_size=args.train_batch_size/2, num_workers=args.num_workers)
+                                  batch_size=int(args.train_batch_size/2), num_workers=args.num_workers)
     global_loss = global_acc = 0
     global_step = 0
     for step, (img_keys, batch) in enumerate(tqdm(train_dataloader)):
